requirements.txt
streamlit==1.28.0
psycopg2-binary==2.9.9
pymongo==4.6.0
pandas==2.1.3
matplotlib==3.8.2